--- centimeter/components/Presentation/PresentationCanvas.vue
@@ -1,11 +1,0 @@
-<template>
-    <div class="h-screen flex items-start justify-center bg-[var(--faded-bg-color-light)] p-6 pt-12">
-        <div
-            role="region"
-            aria-label="Presentation canvas"
-            class="w-[96vw] max-w-[1400px] aspect-video bg-[var(--faded-bg-color)] rounded-[6px] shadow-[0_10px_30px_rgba(0,0,0,0.12)] border border-[rgba(0,0,0,0.06)] overflow-hidden"
-        ></div>
-    </div>
-</template>
-
-<script setup lang="ts"></script>
