+++ centimeter/components/features/dashboard/PresentationCards.vue/CompactPresentationView.vue
