+++ centimeter/components/presentation/editor/EditorBar.vue
