+++ centimeter/components/presentation/editor/NewSlides.vue
