+++ centimeter/components/presentation/editor/SlidesList.vue
