+++ centimeter/components/presentation/panels/EditPanel.vue
