+++ centimeter/components/presentation/slide-components/TextComponent.vue
