+++ centimeter/components/presentation/slide-types/QuizCompletions.vue
