+++ centimeter/pages/auth/login.vue
@@ -6,6 +6,6 @@
 </template>
 
 <script setup lang="ts">
-import NavBar from '~/components/NavBar.vue'
-import LoginPage from '~/components/LoginPage.vue'
+import NavBar from "~/components/Global/NavBar.vue";
+import LoginPage from "~/components/Accounts/LoginPage.vue";
 </script>
