+++ centimeter/pages/auth/signup.vue
@@ -6,6 +6,6 @@
 </template>
 
 <script setup lang="ts">
-import NavBar from '~/components/NavBar.vue'
-import SignUpPage from '~/components/SignUpPage.vue'
+import NavBar from "~/components/Global/NavBar.vue";
+import SignUpPage from "~/components/Accounts/SignUpPage.vue";
 </script>
