+++ centimeter/pages/index.vue
@@ -1,18 +1,15 @@
 <script setup lang="ts">
-import NavBar from '~/components/NavBar.vue';
-import homeindex from '~/components/home-index.vue'
-import JoinBar from '~/components/JoinBar.vue';
-
+import NavBar from "~/components/Global/NavBar.vue";
+import homeindex from "~/components/Home/home-index.vue";
+import JoinBar from "~/components/Home/JoinBar.vue";
 </script>
 
 <template>
   <main>
-    <NavBar/>
-    <JoinBar/>
+    <NavBar />
+    <JoinBar />
   </main>
-  <homeindex/>
+  <homeindex />
 
-  <div>
-    
-  </div>
+  <div></div>
 </template>
